[DSM] DDP-7174: DSM_use_role_migration (#1545)
* DDP-7174
* Changed UserUtil.java and any class that was using it
* Created two Jdbi classes, `JdbiUser.java` and `JdbiUserRole.java` for new Database handling
* Changed `AuthenticationRoute.java` to use the guid from the db for the claims

* added todo

* creating AssigneeDto and removing Assignee.java

* participant list loading working with new db

* Added Daos and Dtos and Jdbi classes for User, UserRoleDao, UserPermissions and UserSettingsDao

Changed other classes to use the shared db

* added code for migrating roles and users

* inserting auth0 user id when missing
changing UserSettingsDao
changing user id from int to long
adding study_guid to ddp_instance queries

* Merging develop

* checkstyle fixes

* checkstyle fixes

* [DSM] DDP-7234: user_changes (#1612)

* DDP-7234 adding users

* DDP-7234 adding user_add permission
DDP-7234 adding new_user permissions

* DDP-7546 `GetRoleRoute` and relevant classes added

* changes to Dtos for json names

* DDP-7235 user access permission

* DDP-7235 assigning new roles to user

* DDP-7238 liquibase

* DDP-7234 changes to match the workflow of thee design

* DDP-7235
*removed new user role insertion query in `ddp-database`
*changed the `modify` functions to assign roles or change `user_profile` info

* DDP-7238
* adding `user_delete` permissions
* adding `DeactivateUserRoute.java` for deletion of user
* changed `User` and `Role` related classes to check for user being active
DDP-7235
*changes to `ModifyUserRoute` class to work with the designs

* DP-7234 Adding `study_admin` liquibase and instance role and role for studies that have phone numbers

* Merging develop

* checkstyle fixes

* checkstyle fixes

* DDP-7234 PR comments to use more Lombok annotations applied

* [DSM] DDP-7883_copy_dsm_user_id (#1641)

* DP-7238 Assignee table changes

* DP-7238 KitType changes to not use old access tables

* DP-7238 `ParticipantExit` changes to not use old access tables and get user names after retrieving exited pts

* DP-7238

* `SurveyTrigger` changes to not use old access tables and get user names after retrieving surveys
* Creating new method to return list of DSM users whether active or not

* DP-7238

* `AbstractionActivity` changes to not use old access tables and get user names after getting abstractions

* DP-7238

* `selectAllowedRealmsStartsLike` never used

* DP-7238

* change `KitDiscard` class for setting the usernames and also the `exitedKits` query in `application.conf`

* DP-7238

* change `ParticipantEvent` class for setting the usernames and also the `selectParticipantEvents` query in `application.conf`

* DDP-7883 removing queries from application.conf to classes

* checkstyle fixes

* DDP-7883 changing order to work with AllArgsConstructor

+++ pepper-apis/dsm-core/src/main/java/org/broadinstitute/dsm/db/EditParticipantMessage.java
@@ -50,13 +50,13 @@ public class EditParticipantMessage {
         this.received_message = received_message;
     }
 
-    public static EditParticipantMessage getMessageWithStatus(int userId) {
+    public static EditParticipantMessage getMessageWithStatus(Long userId) {
         List<EditParticipantMessage> messagesWithStatus = new ArrayList<>();
         SimpleResult results = inTransaction((conn) -> {
             SimpleResult dbVals = new SimpleResult();
 
             try (PreparedStatement stmt = conn.prepareStatement(SQL_SELECT_MESSAGE_AND_STATUS)) {
-                stmt.setInt(1, userId);
+                stmt.setLong(1, userId);
                 try (ResultSet rs = stmt.executeQuery()) {
                     while (rs.next()) {
                         messagesWithStatus.add(
